Refactor packages to reflect business functions
