+++ app/src/main/java/com/vvsoft/saathi/info/record/InfoRecordConfiguration.java
@@ -1,9 +1,13 @@
 package com.vvsoft.saathi.info.record;
 
 import com.vvsoft.saathi.entity.dao.GenericDao;
-import com.vvsoft.saathi.info.GenericLocalStorageDao;
-import com.vvsoft.saathi.info.record.service.InfoRecordCrudService;
-import com.vvsoft.saathi.info.schema.SchemaRepository;
+import com.vvsoft.saathi.info.persistance.GenericLocalStorageDao;
+import com.vvsoft.saathi.info.record.model.InfoRecord;
+import com.vvsoft.saathi.info.record.crud.InfoRecordCrudService;
+import com.vvsoft.saathi.info.record.crud.InfoRecordRepository;
+import com.vvsoft.saathi.info.record.crud.InfoRecordRepositoryGenericImpl;
+import com.vvsoft.saathi.info.record.crud.InfoRecordServiceImpl;
+import com.vvsoft.saathi.info.schema.crud.SchemaRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
