+++ app/src/main/java/com/vvsoft/saathi/info/schema/SchemaConfiguration.java
@@ -1,7 +1,9 @@
 package com.vvsoft.saathi.info.schema;
 
 import com.vvsoft.saathi.entity.dao.GenericDao;
-import com.vvsoft.saathi.info.GenericLocalStorageDao;
+import com.vvsoft.saathi.info.persistance.GenericLocalStorageDao;
+import com.vvsoft.saathi.info.schema.crud.SchemaRepository;
+import com.vvsoft.saathi.info.schema.crud.SchemaRepositoryImpl;
 import com.vvsoft.saathi.info.schema.model.InfoSchema;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
