+++ app/src/main/java/com/vvsoft/saathi/info/schema/crud/api/controller/SchemaErrorHandler.java
@@ -1,8 +1,8 @@
-package com.vvsoft.saathi.info.schema.controller;
+package com.vvsoft.saathi.info.schema.crud.api.controller;
 
 import com.vvsoft.saathi.error.rest.ApiError;
 import com.vvsoft.saathi.error.rest.ApiErrorUtil;
-import com.vvsoft.saathi.info.record.exception.FieldNotFoundInSchemaException;
+import com.vvsoft.saathi.info.record.model.exception.FieldNotFoundInSchemaException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
