+++ app/src/test/java/com/vvsoft/saathi/info/record/InfoRecordServiceImplTest.java
@@ -1,8 +1,10 @@
 package com.vvsoft.saathi.info.record;
 
 import com.vvsoft.saathi.entity.dao.exception.EntityNotFoundException;
-import com.vvsoft.saathi.info.record.dto.InfoRecordDto;
-import com.vvsoft.saathi.info.record.service.InfoRecordCrudService;
+import com.vvsoft.saathi.info.record.presentation.InfoRecordDto;
+import com.vvsoft.saathi.info.record.model.InfoRecord;
+import com.vvsoft.saathi.info.record.model.SimpleRecordValue;
+import com.vvsoft.saathi.info.record.crud.InfoRecordCrudService;
 import com.vvsoft.saathi.info.schema.model.InfoSchema;
 import com.vvsoft.saathi.test.util.StorageUtil;
 import com.vvsoft.saathi.test.util.TestSchemaProvider;
