+++ entity/src/main/java/com/vvsoft/saathi/info/record/model/exception/FieldNotFoundInSchemaException.java
@@ -1,4 +1,4 @@
-package com.vvsoft.saathi.info.record.exception;
+package com.vvsoft.saathi.info.record.model.exception;
 
 import lombok.Getter;
